Restore examples' serialization bench
