+++ rspace-bench/src/test/scala/coop/rchain/rspace/bench/serialization/Base.scala
@@ -1,0 +1,8 @@
+package coop.rchain.rspace.bench.serialization
+
+import java.nio.ByteBuffer
+
+trait Serialize2ByteBuffer[A] {
+  def encode(a: A): ByteBuffer
+  def decode(bytes: ByteBuffer): A
+}
